refactor(core): move flush loop and metrics cluster to dial9-core
