+++ dial9-core/src/flush_loop.rs
@@ -1,14 +1,11 @@
-use crate::metrics::{FlushMetrics, Operation, TlDrainMetrics};
+use crate::buffer;
+use crate::handle::ControlCommand;
+use crate::metrics::{FlushMetrics, FlushStats, Operation, TlDrainMetrics};
 use crate::rate_limit::rate_limited;
-use crate::telemetry::buffer;
-use crate::telemetry::writer::{SegmentWriter, WriterMode};
+use crate::shared_state::SharedState;
+use crate::writer::{SegmentWriter, WriterMode};
 use metrique::timers::Timer;
-use metrique::unit::Microsecond;
-use metrique::unit_of_work::metrics;
 use std::time::Duration;
-
-use super::ControlCommand;
-use super::SharedState;
 
 /// Tracks the drain coordination state between the flush loop and the writer.
 ///
@@ -23,7 +20,7 @@ use super::SharedState;
 /// fires every cycle (since we haven't drained yet), forever deferring the
 /// actual drain.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub(super) enum DrainState {
+enum DrainState {
     /// Normal operation — poll `should_drain()` each cycle.
     Idle,
     /// The writer reported drain due and we bumped the drain epoch.
@@ -31,20 +28,10 @@ pub(super) enum DrainState {
     EpochBumped,
 }
 
-/// Stats returned by flush for metrics publishing.
-#[metrics(subfield, rename_all = "PascalCase")]
-#[derive(Debug)]
-pub(crate) struct FlushStats {
-    pub event_count: u64,
-    pub dropped_batches: u64,
-    #[metrics(unit = Microsecond)]
-    pub cpu_flush_duration: Duration,
-}
-
 /// Perform one flush cycle: drain CPU profilers, drain the collector, write
 /// events to disk, and flush the writer. This is the only code path that
 /// touches the writer, and it runs exclusively on the flush thread.
-pub(super) fn flush_once<M: WriterMode>(
+fn flush_once<M: WriterMode>(
     writer: &mut SegmentWriter<M>,
     events_written: &mut u64,
     shared: &SharedState,
@@ -105,7 +92,8 @@ pub(super) fn flush_once<M: WriterMode>(
 }
 
 /// The flush thread main loop. Extracted so `TelemetryCore::builder` stays readable.
-pub(super) fn run_flush_loop<M: WriterMode>(
+#[doc(hidden)]
+pub fn run_flush_loop<M: WriterMode>(
     control_rx: crate::primitives::sync::mpsc::Receiver<ControlCommand>,
     shared: &SharedState,
     flush_metrics_sink: &metrique_writer::BoxEntrySink,
